apb_intf: Turn parameters into constants
Both address and data signals can be up to 32 bit wide.  Due to the
simple structure of APB, we can assume a reasonably good synthesizer
optimizes away unused parts of these signals, so it does not hurt to
simply fix them at the maximum value.  Besides, this saves problems with
unsupported interface parametrization in some tools and obviates the
need for data and address width converters.